64bit change for ConcurMobile Project

1. Enable 64 bit support settings
2. Fix warnings related to the change.
3. Ignored warnings in files that are not used like OutOfpocketlistviewController.m and in third party libraries like socket.

See merge request !2

+++ iPhone/Classes/Air Booking/AirShopFilteredResultsVC.m
@@ -765,7 +765,7 @@
     return [a count];
 }
 
--(int)numberOfRowsInViolationsSection
+-(NSUInteger)numberOfRowsInViolationsSection
 {
     if ([self.airSummary.canUseTravelPoints boolValue])
     {
@@ -1106,7 +1106,7 @@
     }
 }
 
-- (void)handleSelectInViolationSectionAtRow:(int)row
+- (void)handleSelectInViolationSectionAtRow:(NSInteger)row
 {
     if ([kRowViolationsText isEqualToString:self.rowsInViolationSection[row]])
     {
